requirements.txt
fastapi
uvicorn
pymongo
pymongo[srv]
python-dotenv
motor
pydantic
